+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/presentation/common/rewardDestination/RewardDestinationEstimations.kt
@@ -1,0 +1,8 @@
+package jp.co.soramitsu.feature_staking_impl.presentation.common.rewardDestination
+
+import jp.co.soramitsu.feature_staking_impl.presentation.staking.main.model.RewardEstimation
+
+class RewardDestinationEstimations(
+    val restake: RewardEstimation,
+    val payout: RewardEstimation,
+)
