+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/presentation/common/rewardDestination/RewardDestinationModel.kt
@@ -1,0 +1,11 @@
+package jp.co.soramitsu.feature_staking_impl.presentation.common.rewardDestination
+
+import jp.co.soramitsu.common.address.AddressModel
+
+sealed class RewardDestinationModel {
+
+    object Restake : RewardDestinationModel()
+
+    class Payout(val destination: AddressModel) : RewardDestinationModel()
+}
+
