+++ feature-staking-impl/src/main/java/jp/co/soramitsu/feature_staking_impl/presentation/common/rewardDestination/RewardDestinationProvider.kt
@@ -1,0 +1,84 @@
+package jp.co.soramitsu.feature_staking_impl.presentation.common.rewardDestination
+
+import androidx.lifecycle.MutableLiveData
+import jp.co.soramitsu.common.address.AddressIconGenerator
+import jp.co.soramitsu.common.address.AddressModel
+import jp.co.soramitsu.common.data.network.AppLinksProvider
+import jp.co.soramitsu.common.resources.ResourceManager
+import jp.co.soramitsu.common.utils.Event
+import jp.co.soramitsu.common.view.bottomSheet.list.dynamic.DynamicListBottomSheet
+import jp.co.soramitsu.feature_staking_api.domain.model.StakingAccount
+import jp.co.soramitsu.feature_staking_impl.domain.StakingInteractor
+import jp.co.soramitsu.feature_staking_impl.domain.rewards.DAYS_IN_YEAR
+import jp.co.soramitsu.feature_staking_impl.domain.rewards.RewardCalculator
+import jp.co.soramitsu.feature_staking_impl.presentation.mappers.RewardSuffix
+import jp.co.soramitsu.feature_staking_impl.presentation.mappers.mapPeriodReturnsToRewardEstimation
+import jp.co.soramitsu.feature_wallet_api.domain.model.Asset
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.launch
+import java.math.BigDecimal
+
+class RewardDestinationProvider(
+    private val resourceManager: ResourceManager,
+    private val interactor: StakingInteractor,
+    private val addressIconGenerator: AddressIconGenerator,
+    private val appLinksProvider: AppLinksProvider,
+) : RewardDestinationMixin.Presentation {
+
+    override val rewardReturnsLiveData = MutableLiveData<RewardDestinationEstimations>()
+    override val showDestinationChooserEvent = MutableLiveData<Event<DynamicListBottomSheet.Payload<AddressModel>>>()
+
+    override val rewardDestinationFlow = MutableStateFlow<RewardDestinationModel>(RewardDestinationModel.Restake)
+
+    override val openBrowserEvent = MutableLiveData<Event<String>>()
+
+    override fun payoutClicked(scope: CoroutineScope) {
+        scope.launch {
+            val currentAccount = interactor.getSelectedAccount()
+
+            rewardDestinationFlow.value = RewardDestinationModel.Payout(generateDestinationModel(currentAccount))
+        }
+    }
+
+    override fun payoutTargetClicked(scope: CoroutineScope) {
+        val selectedDestination = rewardDestinationFlow.value as? RewardDestinationModel.Payout ?: return
+
+        scope.launch {
+            val accountsInNetwork = accountsInCurrentNetwork()
+
+            showDestinationChooserEvent.value = Event(DynamicListBottomSheet.Payload(accountsInNetwork, selectedDestination.destination))
+        }
+    }
+
+    override fun payoutDestinationChanged(newDestination: AddressModel) {
+        rewardDestinationFlow.value = RewardDestinationModel.Payout(newDestination)
+    }
+
+    override fun learnMoreClicked() {
+        openBrowserEvent.value = Event(appLinksProvider.payoutsLearnMore)
+    }
+
+    override fun restakeClicked() {
+        rewardDestinationFlow.value = RewardDestinationModel.Restake
+    }
+
+    override suspend fun updateReturns(rewardCalculator: RewardCalculator, asset: Asset, amount: BigDecimal) {
+        val restakeReturns = rewardCalculator.calculateReturns(amount, DAYS_IN_YEAR, true)
+        val payoutReturns = rewardCalculator.calculateReturns(amount, DAYS_IN_YEAR, false)
+
+        val restakeEstimations = mapPeriodReturnsToRewardEstimation(restakeReturns, asset.token, resourceManager, RewardSuffix.APY)
+        val payoutEstimations = mapPeriodReturnsToRewardEstimation(payoutReturns, asset.token, resourceManager, RewardSuffix.APR)
+
+        rewardReturnsLiveData.value = RewardDestinationEstimations(restakeEstimations, payoutEstimations)
+    }
+
+    private suspend fun accountsInCurrentNetwork(): List<AddressModel> {
+        return interactor.getAccountsInCurrentNetwork()
+            .map { generateDestinationModel(it) }
+    }
+
+    private suspend fun generateDestinationModel(account: StakingAccount): AddressModel {
+        return addressIconGenerator.createAddressModel(account.address, AddressIconGenerator.SIZE_MEDIUM, account.name)
+    }
+}
